Seperate reducers and actions to component folders
