+++ src/MainApp.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+import TimerApp from './components/TimerApp/TimerApp'
+
+class MainApp extends Component {
+  render() {
+    return (
+      <TimerApp />
+
+    )
+  }
+}
+
+export default MainApp;
