+++ src/components/RoundsInput/RoundsInputActions.js
@@ -1,0 +1,11 @@
+export const PlusRound = () => {
+    return {
+        type: 'PLUS_ROUND'
+    }
+}
+
+export const MinusRound = () => {
+    return {
+        type: 'MINUS_ROUND'
+    }
+}
