+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import MainApp from './MainApp';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 
@@ -8,6 +8,6 @@ import reducers from './reducers'
 
 ReactDOM.render(
     <Provider store={createStore(reducers)}>
-        <App />
+        <MainApp />
     </Provider>, document.getElementById('root')
 );
